Moving the JVM specific code of the Kotlin runtime under package org.antlr.v5.runtime.kotlin.jvm
