+++ runtime-testsuite/test/org/antlr/v5/test/runtime/kotlin/commonTest/org/antlr/v4/kotlinruntime/CharStreamsTest.kt
@@ -5,7 +5,7 @@
  */
 package org.antlr.v5.test.runtime.kotlin.commonTest.org.antlr.v5.kotlinruntime
 
-import org.antlr.v5.runtime.kotlin.CharStreams
+import org.antlr.v5.runtime.kotlin.jvm.CharStreams
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
