+++ runtime-testsuite/test/org/antlr/v5/test/runtime/kotlin/commonTest/org/antlr/v4/kotlinruntime/StringCharStreamTest.kt
@@ -6,7 +6,7 @@
 package org.antlr.v5.test.runtime.kotlin.commonTest.org.antlr.v5.kotlinruntime
 
 import org.antlr.v5.runtime.kotlin.ext.codePointIndices
-import org.antlr.v5.runtime.kotlin.CharStreams
+import org.antlr.v5.runtime.kotlin.jvm.CharStreams
 import org.antlr.v5.runtime.kotlin.misc.Interval
 import org.junit.jupiter.api.Test
 import kotlin.test.assertContentEquals
