+++ runtime-testsuite/test/org/antlr/v5/test/runtime/states/jvm/KotlinCompiledState.java
@@ -1,6 +1,7 @@
 package org.antlr.v5.test.runtime.states.jvm;
 
 import org.antlr.v5.runtime.kotlin.*;
+import org.antlr.v5.runtime.kotlin.jvm.CharStreams;
 import org.antlr.v5.test.runtime.RuntimeRunner;
 import org.antlr.v5.test.runtime.states.GeneratedState;
