+++ runtime/Kotlin/src/org/antlr/v5/runtime/kotlin/DiagnosticErrorListener.kt
@@ -8,6 +8,7 @@ package org.antlr.v5.runtime.kotlin
 
 import org.antlr.v5.runtime.kotlin.atn.ATNConfigSet
 import org.antlr.v5.runtime.kotlin.dfa.DFA
+import org.antlr.v5.runtime.kotlin.jvm.BitSet
 import org.antlr.v5.runtime.kotlin.misc.Interval
 
 /**
