+++ runtime/Kotlin/src/org/antlr/v5/runtime/kotlin/atn/ATNSimulator.kt
@@ -6,8 +6,8 @@
 
 package org.antlr.v5.runtime.kotlin.atn
 
-import org.antlr.v5.runtime.kotlin.IdentityHashMap
-import org.antlr.v5.runtime.kotlin.synchronized
+import org.antlr.v5.runtime.kotlin.jvm.IdentityHashMap
+import org.antlr.v5.runtime.kotlin.jvm.synchronized
 import org.antlr.v5.runtime.kotlin.dfa.DFA
 import org.antlr.v5.runtime.kotlin.dfa.DFAState
