+++ runtime/Kotlin/src/org/antlr/v5/runtime/kotlin/atn/LL1Analyzer.kt
@@ -6,7 +6,7 @@
 
 package org.antlr.v5.runtime.kotlin.atn
 
-import org.antlr.v5.runtime.kotlin.BitSet
+import org.antlr.v5.runtime.kotlin.jvm.BitSet
 import org.antlr.v5.runtime.kotlin.RuleContext
 import org.antlr.v5.runtime.kotlin.Token
 import org.antlr.v5.runtime.kotlin.misc.IntervalSet
