+++ runtime/Kotlin/src/org/antlr/v5/runtime/kotlin/jvm/BitSet.kt
@@ -3,6 +3,9 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
-package org.antlr.v5.runtime.kotlin
 
-public typealias CopyOnWriteArrayList<E> = java.util.concurrent.CopyOnWriteArrayList<E>
+package org.antlr.v5.runtime.kotlin.jvm
+
+import java.util.BitSet as JavaBitSet
+
+public typealias BitSet = JavaBitSet
