+++ runtime/Kotlin/src/org/antlr/v5/runtime/kotlin/jvm/CopyOnWriteArrayList.kt
@@ -3,9 +3,6 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
+package org.antlr.v5.runtime.kotlin.jvm
 
-package org.antlr.v5.runtime.kotlin
-
-import java.util.IdentityHashMap as JavaIdentityHashMap
-
-public typealias IdentityHashMap<K, V> = JavaIdentityHashMap<K, V>
+public typealias CopyOnWriteArrayList<E> = java.util.concurrent.CopyOnWriteArrayList<E>
