+++ runtime/Kotlin/src/org/antlr/v5/runtime/kotlin/jvm/Synchronized.kt
@@ -3,9 +3,7 @@
  * Use of this file is governed by the BSD 3-clause license that
  * can be found in the LICENSE.txt file in the project root.
  */
+package org.antlr.v5.runtime.kotlin.jvm
 
-package org.antlr.v5.runtime.kotlin
-
-import java.util.WeakHashMap as JavaWeakHashMap
-
-public typealias WeakHashMap<K, V> = JavaWeakHashMap<K, V>
+public inline fun <R> synchronized(lock: Any, block: () -> R): R =
+  kotlin.synchronized(lock, block)
